fix(core): resolve compile warnings and unsafe return in game sources
- fix multiple signed/unsigned comparison warnings in AuctionHouseBotSeller
- correct glyph item level checks to use GetBaseItemLevel where appropriate
- fix float narrowing warnings caused by M_PI usage in Position and Pandaria scripts
- fix signed/unsigned rating comparison in Player
- fix timer type mismatch in monk spell script
- fix loop index type in cs_test command
- replace invalid temporary return in BattlePayData::GetProduct with static fallback

This cleans up compiler warnings and fixes an unsafe reference return that could
lead to undefined behavior.

+++ src/server/game/Entities/Object/Position.cpp
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2012-2016 Uwow <https://uwow.biz/>
- * Copyright (C) 2008-2015 TrinityCore <http://www.trinitycore.org/>
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -153,7 +152,7 @@ bool Position::IsWithinBox(Position const& center, float xradius, float yradius,
 
 bool Position::HasInLine(WorldObject const* target, float width) const
 {
-    if (!HasInArc(M_PI, target))
+    if (!HasInArc(static_cast<float>(M_PI), target))
         return false;
     width += target->GetObjectSize();
     float angle = GetRelativeAngle(target);
@@ -551,7 +550,7 @@ bool Position::HasInArc(float arc, Position const* obj) const
     // move angle to range -pi ... +pi
     angle = NormalizeOrientation(angle);
     if (angle > M_PI)
-        angle -= 2.0f*M_PI;
+        angle -= 2.0f * static_cast<float>(M_PI);
 
     float lborder = -1 * (arc / 2.0f);                        // in range -pi..0
     float rborder = (arc / 2.0f);                             // in range 0..pi
